Move farm update into farm.go
